+++ src/main/java/hobbiedo/crew/kafka/dto/CrewScoreDTO.java
@@ -1,0 +1,8 @@
+package hobbiedo.crew.kafka.dto;
+
+import lombok.Getter;
+
+@Getter
+public class CrewScoreDTO {
+	private long crewId;
+}
